requirements.txt
python-telegram-bot==21.6
openai==2.31.0
redis==5.2.1
python-dotenv==1.0.1
